Add support to pass the target file id as argument of the Hallebarde command
The Hallebarde command can be used with or without parameters. If no
parameter is passed, the v:count variable is used. If a parameter is
passed to the command, this is used as the target file id in the list.

This can be used to bind a shortcut to a particuluar file, as follow:

```
    nnoremap <silent> <Leader><Leader> <CMD>Hallebarde<CR>
    nnoremap <silent> <Leader>a <CMD>Hallebarde(1)<CR>
    nnoremap <silent> <Leader>s <CMD>Hallebarde(2)<CR>
```